// web/src/api/generated/models/PreferenceCategory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI ist der verbindliche Vertrag.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PreferenceCategory = {
    FOOD: 'FOOD',
    DRINK: 'DRINK',
    FLOWERS: 'FLOWERS',
    MOVIES: 'MOVIES',
    SERIES: 'SERIES',
    MUSIC: 'MUSIC',
    HOBBIES: 'HOBBIES',
    ACTIVITIES: 'ACTIVITIES',
    TRAVEL: 'TRAVEL',
    RESTAURANTS: 'RESTAURANTS',
    COLORS: 'COLORS',
    OTHER: 'OTHER'
} as const;
export type PreferenceCategory = typeof PreferenceCategory[keyof typeof PreferenceCategory];


export function instanceOfPreferenceCategory(value: any): boolean {
    for (const key in PreferenceCategory) {
        if (Object.prototype.hasOwnProperty.call(PreferenceCategory, key)) {
            if (PreferenceCategory[key as keyof typeof PreferenceCategory] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PreferenceCategoryFromJSON(json: any): PreferenceCategory {
    return PreferenceCategoryFromJSONTyped(json, false);
}

export function PreferenceCategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): PreferenceCategory {
    return json as PreferenceCategory;
}

export function PreferenceCategoryToJSON(value?: PreferenceCategory | null): any {
    return value as any;
}

export function PreferenceCategoryToJSONTyped(value: any, ignoreDiscriminator: boolean): PreferenceCategory {
    return value as PreferenceCategory;
}